src/requirements.txt
mysql-connector-python>=8.0.0
confluent-kafka>=2.2.0
python-dotenv>=1.0.0
